requirements.txt
flatten_json==0.1.13
pandas==1.5.3
python_dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
snowflake-connector-python==3.0.0
snowflake_sqlalchemy==1.4.6
SQLAlchemy==1.4.46
tabulate==0.9.0
pymssql==2.2.7
